Put the rule semantics specs under Formatting.Rules
Precedence and combining two rulesets are properties of RuleSet, not of the
formatter; the formatter was only the instrument they were observed through.
Adds the two that were missing: a prefix key reaches below the node it names,
and length settles before specificity. matchModeSpec now builds its rules from
JBFL source too, so nothing here constructs a RuleSet by hand.
